harden(sim): encapsulate inventory cache, assert invariants, add UBSan CI (#483)
* ci: add UBSan + ASan post-deploy verification job

Adds test-ubsan to the post-deploy gate (needs: deploy, main-only),
mirroring test-asan but with -fsanitize=undefined,address. UBSan catches
signed overflow, strict-aliasing, null derefs, and OOB that ASan alone
misses. UBSAN_OPTIONS=halt_on_error=1:print_stacktrace=1 and
ASAN_OPTIONS=halt_on_error=1 ensure first-failure surfaces with a stack.

Fix: spatial_grid hash multiplied signed cx by 73856093, overflowing for
|cx| > 29 (UB caught by UBSan). Multiplication now happens in unsigned
space, which matches the existing uint32_t result type and is well-defined.
Both call sites (game_sim.c and game_sim.h inline lookup) updated.


* sim: assert manifest/inventory invariant in finished-good accessors

Add NDEBUG-gated assert_finished_invariant() at the end of each
station_finished_{mint,drain,accumulate}: floor(inventory[c]+eps) must
equal manifest_count_by_commodity(c). Also assert manifest->count <=
manifest->cap at the end of manifest_push and manifest_reserve.

Wiring up the assertion exposed a real bug in mint's float-sync path:
when accumulate's int_after = floor(after + 0.0001f) rolls over from N
to N+1 with after = 0.99999..., mint would compute frac = inv -
floorf(inv) = 0.99999..., then set inventory[c] = manifest_count + 0.99999,
which lands one full unit above the manifest count and breaks the
invariant. Fix: apply the same +0.0001 epsilon to the floor in mint's
sync, matching accumulate's rounding contract.


* sim: rename station_t::inventory to _inventory_cache

The float array on station_t is the cache layer of the manifest for
finished goods (c >= COMMODITY_RAW_ORE_COUNT) — the manifest is the
source of truth, the float is a derived count. Direct writes to
finished-good slots silently break the floor(inventory[c]) ==
manifest_count_by_commodity(c) invariant; raw-ore slots (the smelter
hopper) are float-only and may still be written directly.

Renaming the field to _inventory_cache (leading underscore = private)
forces every existing read/write to surface in code review and gives
a one-stop docstring pointing future writers at the
station_finished_{mint,drain,accumulate} accessors. The wire format is
unchanged: read_station / write_station_session in sim_save.c serialize
the float array as raw bytes — the layout doesn't depend on the C name.

Save tests (test_save_file_size_stable, test_save_load_preserves_*,
v21 remap) all pass; manifest invariant assertions from the prior
commit pass at -O0/-Og/-O2 and under UBSan+ASan.

This commit deliberately stops at the rename — converting every direct
write to an accessor is a larger behavioral change (tonnes of float-only
test fixtures rely on the existing reconciler in world_sim_step that
snaps drift up to manifest count). Those conversions can be done
incrementally per-site now that the field name signals intent.


---------

+++ server/game_sim.h
@@ -109,7 +109,8 @@ static inline void spatial_grid_cell(const spatial_grid_t *g, vec2 pos, int *cx,
 /* Look up a cell by coordinates. Returns NULL if empty. */
 static inline const spatial_cell_t *spatial_grid_lookup(const spatial_grid_t *g, int cx, int cy) {
     if (!g->entries) return NULL;
-    uint32_t h = (uint32_t)((cx * 73856093) ^ (cy * 19349663));
+    /* Mul in unsigned space — signed * 73856093 overflows for |cx| > 29 (UB). */
+    uint32_t h = ((uint32_t)cx * 73856093u) ^ ((uint32_t)cy * 19349663u);
     for (uint32_t i = h & g->mask; ; i = (i + 1) & g->mask) {
         const sparse_cell_entry_t *e = &g->entries[i];
         if (e->key_x == INT32_MIN) return NULL;      /* empty slot */
